% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igmi_barycenter.R
\name{bw_dist2}
\alias{bw_dist2}
\title{Squared 2-Wasserstein distance between Gaussian study objects}
\usage{
bw_dist2(N1, N2)
}
\arguments{
\item{N1, N2}{\code{igmi_gaussian} objects.}
}
\value{
Nonnegative scalar.
}
\description{
\verb{W_2^2 = ||theta_1 - theta_2||^2 + B^2(Sigma_1, Sigma_2)}
(splitting, igmi_notes Thm 2.4).
}
