% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igmi_inference.R
\name{frechet_scatter}
\alias{frechet_scatter}
\title{Frechet scatter of the study effects}
\usage{
frechet_scatter(studies, weights = NULL)
}
\arguments{
\item{studies}{List of \code{igmi_gaussian} objects.}

\item{weights}{Positive weights (default equal).}
}
\value{
m x m matrix.
}
\description{
\verb{Omega_hat = sum_i w_i (theta_i - theta_bar)(theta_i - theta_bar)^T}
with \verb{tr Omega_hat = V_loc} (igmi_notes Prop 5.16).
}
