% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igmi_gaussian.R
\name{igmi_studies}
\alias{igmi_studies}
\title{Construct a list of IGMI study objects}
\usage{
igmi_studies(yi = NULL, sei = NULL, theta = NULL, Sigma = NULL)
}
\arguments{
\item{yi, sei}{Univariate effect estimates and standard errors.}

\item{theta}{K x m matrix of multivariate estimates.}

\item{Sigma}{List of K m x m covariances.}
}
\value{
List of \code{igmi_gaussian} objects.
}
\description{
Univariate: pass \code{yi} and \code{sei}. Multivariate: pass \code{theta} (K x m
matrix) and \code{Sigma} (list of K covariance matrices) -- the output
format of \code{\link[=sim_multivariate]{sim_multivariate()}} and \code{\link[=igmi_study_list]{igmi_study_list()}}.
}
