% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric_mse}
\alias{metric_mse}
\title{Mean squared error over replications}
\usage{
metric_mse(est, truth)
}
\arguments{
\item{est}{Estimates.}

\item{truth}{True value (scalar or vector like \code{est}).}
}
\description{
Mean squared error over replications
}
